Finalized core Git support
What I don't like about this approach is that it messes with the user's
system by calling `git stash`. I should try to find a safer and
less intrusive approach.
On the upside, this approach makes it really easy to go back in time and
analyze the committed codebase. I really like the block syntax
used in the travel_to_head method, which mimics Timecop.