Implement abstract `root_directory` method in ViewHelpers
From Practical Object-Oriented Design In Ruby, page 151, by Sandi Metz:

"Even if there were no reasonable application default for [the foo
method], the [bar] module must still implement the [foo] method. The
rules for modules are the same as for classical inheritance. If a module
sends a message it must provide an implementation, even if that
implementation merely raises an error indicating that users of the
module must implement the method."