Create "Churn vs Complexity" scatter plot
Probably the most useful feature of this gem. The problems I'm feeling
while developing this gem can already be gleaned just by analysing the
scatter plot of this gem's codebase:

* The Reporter class is getting unmanageable.

* A SourceControlSystem instance is always being passed around from
class to class. Should it be some kind of global singleton object? I'm
tempted to put it inside the Configuration class, but that doesn't feel
right.