Disable UtilityFunction smell
Not only do I disagree a bit with this smell but it was also being
reported every time a FeatureEnvy smell was detected.

These smells are so coupled I couldn't even create a test for this
feature. There's probably a way though.

In my opinion, classes should be allowed to have some helper methods.
Not everything has to be an object.