Refactor system
The point of this refactor was to to create a new class, AnalysedFile,
to wrap a file's smells, churn and complexity. In the future, this class
will also hold a file's rating.

The Turbulence class changed a lot in this refactor. It now merely
receives a bunch of analysed_files and asks them for their churn and
complexity values. And now it already returns json instead of a hash.

Potential problems:

* I'm thinking in terms of files, not classes

* The AnalysedFilesBuilder#file_smells method is slow