Fix some table columns being larger than others (for no apparent reason)
This is achieved by setting the [`table-layout` CSS property][1]. With this
property, table and column widths are set by the widths of table and col
elements or by the width of the first row of cells. This has the
drawback that subsequent cell content may not fit in the column widths
provided.

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/table-layout