Add right padding to the index tables
I don't like the fact that I'm adding a padding to each cell
individually. Ideally, I'd just pad the table, but if I do that, then
the header line separator will be a little shorter. I like the way the
line separator is set up at the moment, so this right padding will do
for now.