Analyse only the files whose paths are specified in the CLI
Previously, RubyCritic would analyse **all** the files so that if the
user switched the analysed directories (for instance, from "app" to
"lib") then the revision file serialized would already contain the "lib"
files analysed. This was important because otherwise "old" smells in
"lib" files would erroneously be marked as "new".

Now, RubyCritic only analyses the files specified in the CLI and uses
the paths to all the files as a cache key in the revision file path.

Still, the Orchestrator and RevisionComparator classes are looking
horrible, wish I knew how to organize all this a little better.