Require Reek 2.0.2.
This version has some small bug fixes, and some features that
don't matter to RubyCritic. The next version will probably fix
a crashing bug when `super` is called with parameters and a block
(troessner/reek#404). This is the configuration change that needs
to happen to upgrade to that version, when it's available.