Respect the .reek configuration file.
Rubycritic was using a proxy from Reek::Examiner.
The examiner does not search (or load) the config files by itself,
so we need to allow the default behavior used by the Reek::CLI::Application
which is either passing a path or a nil, that will trigger the searching for .reek files.