Repaired RuboCop offences without breaking Ruby <2.3 compatibility.
RuboCop’s default behaviour when a string literal has its `#freeze` method called in the presence of a `# frozen_string_literal: true` pragma comment is to flag it as a `Style/RedundantFreeze` violation. Removing the `#freeze` calls *should* silence RuboCop under Ruby 2.3 and later, but causes `Style/MutableConstant` violations to be flagged instead. Disabling the `Style/RedundantFreeze` cop for the affected source files in the `.rubocop.yml` file should solve the problem.

[Closes #218]