Traverse git history only once
Before, we were traversing the history for each path.
That behavior gets costly with a long history.

Now, we are traversing the history once initially,
creating a cache from paths to revision counts,
and referencing the cache for each path.

We might be able to do similar for the other source control systems.
However, I personally don't have enough experience with them,
nor do I have an actual project to test on.