Use git commands only
There were two problems with the previous idea:

1. Dropping the `--follow` means we lose information if a file was moved.
2. `sort` and `uniq` will not work well on different platforms.

To address the second, we can use git commands only.
This way, it should have the same behavior on different platforms.

To address the first, we ask for the history of all files ever.
This gives us a pretty big cost up front,
but we cache the results of the command and query them later.
The cost is a bit more than the cost of using
`--follow` once on an old file that has been committed to many times.
But we only have to do it once, rather than for each file.

Once we have the results of the command,
we map them into operations (e.g. `Add` or `Modify`),
then reduce the operations to our cached results for each file name.

Overall, the method should provide for a much more efficient version
than using `--follow` on each file.
Though it will be slightly less efficient than not following renames.