Fix YAML configuration file 'formats' processing
Fixes: Ignores .rubycritic.yml file #303
https://github.com/whitesmith/rubycritic/issues/303.

Update the documentation to indicate that the key in the
.rubycritic.yml configuration file which controls the output
format is 'formats' rather than 'format'.

Fix a bug in RubyCritic::Cli::Options#to_h which prevents a
setting in the YAML configuration file from being used when the
setting is an Array (as is the case for the output formats).

Modify RubyCritic::CLI::Options::File#formats to return an array
of symbols instead of an array of strings, to be consistent with
how RubyCritic::CLI::Options::Argv#parse returns the array. Also
modify it so it can accept a single output format or a list of
formats. The documentation indicates that either is acceptable
but the code assumed the input was always an array.

Signed-off-by: Christopher Voltz <christopher.voltz@hpe.com>