Add Coverage analyser to the project
Analyser::Coverage will analyse test coverage based on a .resultset.json file.

We can assume that projects which are using SimpleCov will have a coverage folder with results. RubyCritic can use the results from that folder to provide a coverage section.

This will help us solve the problem described in https://github.com/whitesmith/rubycritic/issues/245.

It also addresses an item in the ROADMAP.md file:

```
 Integrate other gems, like:

- Simplecov to provide code coverage
```

Analyser won't do anything if a .resultset.json is not present in the caller's coverage directory.