Add an HTML generator for SimpleCov
It adds a new section to the HTML report: Coverage.

In that section RubyCritic will now list all the application files along with its test coverage information.

This change will be necessary to display the results calculated by Analyser::Coverage and it will help improve the situation for this issue: https://github.com/whitesmith/rubycritic/issues/245