Take .flayignore file into account
Flay accepts a `.flayignore` file containing paths allowing to skip its
analysis for the provided paths. The current version of Rubycritic
ignores this file and Flay will always apply to all the paths provided
to Rubycritic.

This patch addresses this issue by taking a small snippet of code from
the Flay source itself and apply it to the Flay helper class. Now the
`.flayignore` file is honored as expected.