Allow the Rake task to accept a custom description
I want to have multiple Rake tasks in my app: one for local mode and one for ci mode.
Unfortunately, the Rake task class doesn't accept a description, so they both display "Run RubyCritic" in the output of `rake -T`.

This fixes that problem by adding the description as a param to the initializer.

Co-Authored-By: Andy Nicholson <andrew@anicholson.net>