Bugfix: Exit with 0 or 1 from Rake Task (#399)
* Exit with 0 or 1 from Rake Task. Fixes #214

Expect the Rake task to behave in the same manner as the cli.
This makes it easier to handle in a CI context. With the
previous behavior, you had to write your own Rake task that
runs the cli as a shell command.
```
task :rubycritic do
  desc 'Run rubycritic and return an exit code'
    sh 'rubycritic --no-browser --minimum-score 95'
end
```
Icky.

* Ignore TooManyInstanceVariables for RakeTask

Public instance variables are necessary for a Rake task. As
configuration options grow, public instance variables grow too. It's
annoying to uptick `max_instance_variables` for each new
configuration option. Instead, ignore that rule for the whole class.

Co-authored-by: Nuno Silva <nunosilva800@gmail.com>