Scope churn calculation to provided paths
The churn calculation was hard-coded to '.'
This meant that rubycritic would run a `git log`
command for the entire repository regardless of
the paths specified with running the `rubycritic`
command.

Running `git log` for large codebases with many
years of commits can take over 30 minutes. This
is particularly wasteful in mono-repositories with
distinct projects that might be better analyzed
separately (large parts of the repository might
not even be Ruby code!).

The time to generate a report decreases significantly
when scoping churn to the paths provided to the
`rubycritic` command.

Other things to note:
- Git log's --follow option can only handle a single
  path at a time. Therefore we need to run multiple
  git log commands when multiple path arguments
  are provided. The results are still merged together.

- The initializer for SourceControlSystem::Git::Churn
  was already at the maximum number of statements
  allowed. I extracted `renames` into a memoized
  method to make room for the new parameter.

- Added tests for SourceControlSystem::Git#churn
  to ensure SourceControlSystem::Git::Churn is
  instantiated with the correct options. This also
  adds a bit of coverage on Config.paths needing
  to exist.