Fix feature execution on Ruby 3.5.0dev
cucumber-core fails on Ruby 3.5.0dev due to a change in the expected method signature for Cucumber::Core::Test::Location#from_source_location. Work around this by aliasing out the old method and then using a new proxy method to catch the method and send the old style message to the old method.

# Conflicts:
#	CHANGELOG.md